backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
yfinance
ta
joblib
pydantic
websockets
python-dotenv
